# backend-admin/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "version": "1.0",
    "description": "API для управления товарами в интернет-магазине"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Локальный сервер"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "summary": "Получить список товаров",
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Добавить новый товар",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товар добавлен"
          }
        }
      }
    },
    "/api/products/{id}": {
      "put": {
        "summary": "Обновить товар",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар обновлен"
          }
        }
      },
      "delete": {
        "summary": "Удалить товар",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар удален"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "description": { "type": "string" },
          "categories": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "number" },
          "description": { "type": "string" },
          "categories": { "type": "array", "items": { "type": "string" } }
        }
      }
    }
  }
}